Move pokerb file to html folder
