+++ html/pokerb/player.rb
