+++ html/pokerb/player_test.rb
